+++ scg-system/src/test/java/io/telicent/AbstractBearerAuthTests.java
@@ -54,6 +54,7 @@ public abstract class AbstractBearerAuthTests {
                 // Ignore
             }
         });
+        Labels.rocks.clear();
     }
 
     @Test